Release/decimal inventory qty (#25)
* feat: decimal quantity for purchase, sales and wastage

Quantity was an integer end to end (INTEGER column, Go int binding,
<input min=1>, zod min(1)), so fractional amounts like 1.25 kg couldn't
be recorded. Rate already accepted 2 decimals (stored as paisa).

- migration 000047: stock_in/stock_out/wastage qty INTEGER -> NUMERIC(12,3)
- sqlc: map the qty columns to float64; handlers round qty to 3 dp
  (utils.RoundQty) before storing
- summary: qty sums are NUMERIC(14,3) instead of ::INTEGER (which would
  truncate); amounts round each line to whole paisa, SUM(ROUND(qty*rate)),
  matching how the supplier ledger records each line
- supplier ledger credit computed from the stored qty/rate with integer
  rounding (utils.LineAmount) instead of raw floats
- reject purchase lines worth < 1 paisa with a 400: they'd otherwise hit
  supplier_ledger's amount > 0 CHECK and 500
- rate bounds: min 0.01 (0.001 rounded to 0 paisa and 500'd on the rate
  CHECK) and max 999999.99 (fits the paisa int32 and NUMERIC(10,2) rate
  columns)
- web: qty inputs accept 3 decimals; shared zod quantity/rate schemas
  with decimal-place limits; the line-items total rounds per line like
  the backend; qty input widened to fit longer decimals

* fix: inventory edit bugs (sale date wipe, purchase edit 500, edit prefill)

- sale edit wiped the date: UpdateStockOut never passed Date, so every
  edit stored date = ''.
- purchase edit always 500'd: UpdateStockIn never set supplier_id (NOT
  NULL). The edit form now has a supplier picker (pre-filled) and the
  handler runs in a transaction that also syncs the purchase's
  auto-recorded supplier ledger credit (supplier, amount via
  utils.LineAmount, BS + AD date via new utils.BSToAD, description).
  Purchases with no credit (recorded before the ledger was wired up) are
  updated without backfilling one. Unknown stock/supplier/product -> 404,
  invalid BS date -> 400.
- purchase delete now removes its auto credit in the same transaction
  instead of leaving it orphaned via ON DELETE SET NULL.
- the auto credit is matched by stock_in_id + entry_type 'cr' +
  created_at = stock_in.created_at (inserted in the same tx), so manual
  ledger entries linked to a purchase are never touched.
- edit dialogs didn't pre-fill the product: the API sends productId but
  the shared record schemas/dialogs used productID. Schemas now mirror
  the API (plus supplierId/supplierName on purchase rows).
- purchase update schema validates the date as BS (the AD regex rejected
  day 32).

* fix: validate BS dates on sales/wastage and use the date picker on purchase edit

- backend: sale and wastage create/update (and purchase update) dates now
  use the existing bs_date validator (go-bs), so a nonexistent BS date
  like 2083-02-32 or free text gets a 400 with a date field error instead
  of being stored as-is.
- web: sale/wastage create and edit schemas validated their BS dates with
  an AD-shaped regex that capped days at 31, rejecting real BS dates like
  2083-03-32. They now use bsDateSchema, which is tightened to months
  01-12 and days 01-32 (the real calendar check stays on the backend).
- purchase edit: the date field is now the same NepaliDatePicker as the
  other forms instead of a plain text box.

* chore: normalize stock_out/wastage rate columns to INTEGER

Migration 000007 declares stock_out.rate and wastage.rate as INTEGER
(paisa), but databases built from an earlier draft of it (dev Neon, and
local copies seeded from it) have NUMERIC(10,2). Migration 000048 brings
them in line with ROUND(rate)::INTEGER: lossless, since the app only ever
writes whole paisa, and effectively a no-op on databases already INTEGER.

The down migration is intentionally a no-op: the pre-up type differs per
database, so there's no single inverse, and converting back to NUMERIC
would introduce drift where there was none. sqlc/app code unchanged (the
schema file already says INTEGER).

+++ apps/backend/internal/database/schema/inventory.sql
@@ -14,7 +14,7 @@ CREATE TABLE stock_in (
     product_id UUID NOT NULL REFERENCES products(id) ON DELETE RESTRICT,
     date TEXT NOT NULL,
     invoice_no TEXT,                 
-    qty INTEGER NOT NULL CHECK (qty > 0),
+    qty NUMERIC(12,3) NOT NULL CHECK (qty > 0),
     rate INTEGER NOT NULL CHECK (rate > 0),
     note TEXT,
     supplier_id UUID NOT NULL REFERENCES suppliers(id) ON DELETE RESTRICT,
@@ -30,7 +30,7 @@ CREATE TABLE stock_out (
     product_id UUID NOT NULL REFERENCES products(id) ON DELETE RESTRICT,
     date TEXT NOT NULL,              
     bill_no TEXT,                
-    qty INTEGER NOT NULL CHECK (qty > 0),
+    qty NUMERIC(12,3) NOT NULL CHECK (qty > 0),
     rate INTEGER NOT NULL CHECK (rate > 0),
     note TEXT,
     created_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
@@ -45,7 +45,7 @@ CREATE TABLE wastage (
     id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
     product_id UUID NOT NULL REFERENCES products(id) ON DELETE RESTRICT,
     date TEXT NOT NULL,              -- BS date string
-    qty INTEGER NOT NULL CHECK (qty > 0),
+    qty NUMERIC(12,3) NOT NULL CHECK (qty > 0),
     rate INTEGER NOT NULL CHECK (rate > 0),
     reason TEXT,
     created_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
